:construction: Restructure big files
If an entity file was excessively large, I just decided to split it up for easier navigation through the code
