+++ core/src/db/entity/book_club/mod.rs
@@ -1,0 +1,12 @@
+mod chat_board;
+mod club;
+mod common;
+mod member;
+mod prisma_macros;
+mod schedule;
+
+pub use chat_board::*;
+pub use club::*;
+pub use member::*;
+pub use prisma_macros::*;
+pub use schedule::*;
