+++ core/src/db/entity/media/mod.rs
@@ -1,0 +1,5 @@
+mod entity;
+mod progress;
+
+pub use entity::*;
+pub use progress::*;
