+++ core/src/db/entity/metadata/mod.rs
@@ -1,0 +1,6 @@
+mod common;
+mod media_metadata;
+mod series_metadata;
+
+pub use media_metadata::*;
+pub use series_metadata::*;
