+++ core/src/db/entity/mod.rs
@@ -1,18 +1,15 @@
-// TODO: fix all these pub uses
-
-pub mod book_club;
+mod book_club;
 pub(crate) mod common;
-pub mod epub;
-pub mod library;
-pub mod log;
-pub mod media;
-pub mod metadata;
-pub mod read_progress;
-pub mod reading_list;
-pub mod series;
-pub mod server_config;
-pub mod tag;
-pub mod user;
+mod epub;
+mod library;
+mod log;
+mod media;
+mod metadata;
+mod reading_list;
+mod series;
+mod server_config;
+mod tag;
+mod user;
 
 pub use self::epub::*;
 pub use self::log::*;
@@ -20,11 +17,10 @@ pub use self::log::*;
 pub use book_club::*;
 pub use library::*;
 pub use media::*;
-pub use read_progress::*;
+pub use metadata::*;
 pub use reading_list::*;
 pub use series::*;
 pub use tag::*;
 pub use user::*;
 
-pub(crate) use common::Cursor;
 pub use common::{FileStatus, LayoutMode};
