+++ core/src/db/entity/series/mod.rs
@@ -1,0 +1,3 @@
+mod entity;
+
+pub use entity::*;
