+++ core/src/db/mod.rs
@@ -11,3 +11,4 @@ pub use client::{create_client, create_client_with_url, create_test_client};
 pub use common::{
 	CountQueryReturn, DBPragma, JournalMode, JournalModeQueryResult, PrismaCountTrait,
 };
+pub use entity::FileStatus;
